requirements.txt
arrow==0.10.0
google-cloud==0.22.0
numpy==1.12.0
scipy==0.18.1
scikit-learn==0.18.1
nltk==3.2.2
urllib3==1.20
matplotlib==2.0.0
selenium==3.0.2
elementium==1.1.7
certifi==2017.1.23
tinydb==3.2.2
pyvirtualdisplay==0.2.1
boto3==1.4.4
psutil==5.2.0
sh==1.12.11
